Move parseIdNumber from client/ -> urbit/
