+++ packages/api/src/urbit/channel.ts
@@ -3,7 +3,6 @@ import bigInt, { BigInteger } from 'big-integer';
 import _ from 'lodash';
 import BTree from 'sorted-btree';
 
-import { parseIdNumber } from '../client/apiUtils';
 import type { Stringified } from '../lib/utilityTypes';
 import {
   Block,
@@ -16,6 +15,7 @@ import {
 } from './content';
 import { Flag } from './hark';
 import { Metadata } from './meta';
+import { parseIdNumber } from './utils';
 
 export interface CacheId {
   author: string;
