+++ packages/api/src/urbit/utils.ts
@@ -1,4 +1,4 @@
-import { da, render, valid } from '@urbit/aura';
+import { da, render, tryParse, valid } from '@urbit/aura';
 import { Atom } from '@urbit/nockjs';
 
 import { PostContent } from '../client';
@@ -343,4 +343,15 @@ export function createSectionId() {
   const newSectionId = `z${idParts[idParts.length - 1]}`;
 
   return newSectionId;
+}
+
+//  parses either a @ud-formatted string (dot-separated decimal) or a plain
+//  decimal number string (aka @ui without the prefix).
+//  we try both instead of being precise at the callsites, because historically
+//  we used a too-lenient @ud parser, which also accepted dotless
+//  representations, making it slightly unclear/ambiguous what we were actually
+//  *intending* to parse. the backend is wildly inconsistent, so this is easier
+//  and safer than figuring all that out. (we'll tighten things up Soon™.)
+export function parseIdNumber(id: string): bigint {
+  return tryParse('ud', id) || BigInt(id);
 }
